lcd_touch: Update FreeRTOS critical sections to use port*_CRITICAL macros instead of task*_CRITICAL macros
Currently the esp_lcd_touch_xx files use the taskENTER_CRITICAL and
taskEXIT_CRITICAL macros to establish a critical section. They use the
macros with a local spinlock variable. This is compatible with existing
FreeRTOS in ESP-IDF. However, the same macros do not accept a spinlock
variable when Amazon FreeRTOS-SMP is used instead of IDF-FreeRTOS.
Therefore, to maintain compatibility across both kernel versions, this
commit updates the critical sections to use the portENTER_CRITICAL and
portEXIT_CRITICAL macros which provide the flexibility to pass local
spinlock variables irrespective of the FreeRTOS kernel type used in
ESP-IDF.